Update doxygen exclude patterns
The previous patterns excluded all files if one of the excluded
directories appeared in the path of the FEAT3 repository itself. Take
for example a FEAT3 repository located in /data/feat3. For that
repository the */data/* exclude pattern would exclude all files in the
repository, not just those in the /data/feat3/data subdirectory.

This commit also adds the out/ subdirectory created by Visual Studio
Code to the excluded patterns.