Fix topology creation in MeshPart operations
Previously a topology for the result mesh part was created any time
either of the input mesh parts had a topology. This behaviour is wrong
for the union of two mesh parts, where the result mesh part covers the
domain of both mesh parts and thus both input mesh parts must have a
topology, so that an input toplogy exists across the whole domain.