major Global::AlgDofParti and Solver::ADPSolverBase overhaul
This commit includes a full rewrite of the algebraic dof partitioning classes
along with the ADPSolverBase class, which is used as a base class for solver
wrappers that enable us to use MPI-parallel third-party solver libraries such
as HYPRE (including BoomerAMG), SuperLU, Nvidia cuDSS, Intel MKL DSS/CSS and
(upcoming) FROSch, which all expect the matrix to be given in a row-wise
partitioned global CSR format -- or just "RPG-CSR" in short -- which is exactly
what the algebraic dof partitioning mechanism of FEAT3 provides.

The major improvement of this commit is that it is now also possible to convert
meta matrix systems (such as e.g. the saddle point matrices in Navier-Stokes
simulations) directly into RPG-CSR format without the need to "scalarize" the
system beforehand, because previously the ADP mechanism could only be applied
to scalar systems discretized into a standard scalar CSR matrix. Furthermore,
this commit also reduces the memory consumption of the solver wrapper classes
by removing the need to store an additional copy of the row-pointer and
column-index arrays for the sake of data-type and index-type conversions.
Now, the ADP classes directly assemble the RPG-CSR matrix arrays, which are
allocated and managed by the third-party solver wrapper classes.

Both the HYPRE and SuperLU wrappers have been updated to work with the new
ADPSolverBase class, therefore enabling these solvers to be applied onto meta
systems, which was not possible before. New wrappers for Nvidia cuDSS and
Intel MKL DSS/CSS are already in the pipeline, which are going to be combined
into a single DirectSparseSolver wrapper class in near future.

Currently, the ADPSolverBase class only supports scalar/blocked unit filters for
the imposition of boundary conditions, so neither the scalar/blocked mean filter
nor the slip filter are currently supported yet.

Signed-off-by: Peter Zajac <peter.zajac@math.tu-dortmund.de>