Fix bug in neighbor calculation of adaptive meshes
The neighbor calculation works by first working out which to elements
are adjacent to a facet of the mesh, and then choosing the respective
other element as the neighbor in that direction.

Up to now the tracking of adjacent elements used a single secondary map,
which silently overwrote some data. The tracking now uses secondary maps
separated by layer and transitivity of the mesh entity, just like the
main storage.

This commit also adds an assertion is the secondary map to catch cases
like these easier in the future.