Add free surface FE assembly class
This commit introduces an assembler base class as well as basic jobs
to allow to integrate a FE function across a freely defined
piecewise constant submanifold of our FE domain.

Note that finding the actual elements to integrate over has a complexity
of O(m*n), where m is the number of mesh elements and n the number of
surface elements. So if possible, use trace assembler on actual mesh
parts, if possible.

Signed-off-by: Maximilian Esser <maximilian.esser@math.tu-dortmund.de>